Update INA219_WE.cpp
According to the datasheet reading the INA219_BUS_REG does not clear the CNVR flag. Reading the INA219_PWR_REG instead would clear the CNVR flag.
Since the INA219_CONF_REG is written afterwards anyways, even a read of the INA219_PWR_REG would be unnecessary.
Checked the behaviour with the Triggerd example and everything seems to be fine.
Thank you for your efforts!